Intento de SprintSecurity 1

+++ src/main/java/pe/edu/upc/connection2connection/repositories/IRoleRepository.java
@@ -5,6 +5,8 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.connection2connection.entities.Role;
 
 @Repository
-public interface IRoleRepository extends JpaRepository<Role, Integer> {
+
+public interface IRoleRepository extends JpaRepository<Role, Long> {
 
 }
+
